Interlink; Rimpoint; Foxhole & Condos: Oh My!
ALSO ADDS MOST OF THE BACKEND FOR LATEJOIN LOCATION STUFF. AND RUINS

good fucking god. if atlas shrugged i have scoliosis.

idk if we're actually going to keep the interlink/condos with the lack of mecherp. the replacement i thought about working on is a year's worth of work i'd rather pour into slashco 13.
that's ultimately not my call ig.

next up: killing the broadcast team.